requirements.txt
Django~=3.0.5
djangorestframework~=3.11.0
djangorestframework_simplejwt~=4.4.0
python-dateutil~=2.8.1
requests~=2.23.0
